backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database dependencies
sqlalchemy>=2.0.23
alembic>=1.12.0

# File processing dependencies
aiofiles>=23.2.0
PyPDF2>=3.0.1
python-docx>=0.8.11

# ML and AI dependencies (optional)
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Optional speech recognition
SpeechRecognition>=3.10.0

# Gemini AI integration
google-generativeai>=0.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
